Modification of Errors

+++ app/Http/Controllers/Omega/CashController.php
@@ -130,18 +130,18 @@ class CashController extends Controller
             }
             $cash->excacc = $excAcc->idaccount;
 
-            $cash->mon1 += trimOver(Request::input('B1'), ' ');
-            $cash->mon2 += trimOver(Request::input('B2'), ' ');
-            $cash->mon3 += trimOver(Request::input('B3'), ' ');
-            $cash->mon4 += trimOver(Request::input('B4'), ' ');
-            $cash->mon5 += trimOver(Request::input('B5'), ' ');
-            $cash->mon6 += trimOver(Request::input('P1'), ' ');
-            $cash->mon7 += trimOver(Request::input('P2'), ' ');
-            $cash->mon8 += trimOver(Request::input('P3'), ' ');
-            $cash->mon9 += trimOver(Request::input('P4'), ' ');
-            $cash->mon10 += trimOver(Request::input('P5'), ' ');
-            $cash->mon11 += trimOver(Request::input('P6'), ' ');
-            $cash->mon12 += trimOver(Request::input('P7'), ' ');
+            $cash->mon1 += (int)trimOver(Request::input('B1'), ' ');
+            $cash->mon2 += (int)trimOver(Request::input('B2'), ' ');
+            $cash->mon3 += (int)trimOver(Request::input('B3'), ' ');
+            $cash->mon4 += (int)trimOver(Request::input('B4'), ' ');
+            $cash->mon5 += (int)trimOver(Request::input('B5'), ' ');
+            $cash->mon6 += (int)trimOver(Request::input('P1'), ' ');
+            $cash->mon7 += (int)trimOver(Request::input('P2'), ' ');
+            $cash->mon8 += (int)trimOver(Request::input('P3'), ' ');
+            $cash->mon9 += (int)trimOver(Request::input('P4'), ' ');
+            $cash->mon10 += (int)trimOver(Request::input('P5'), ' ');
+            $cash->mon11 += (int)trimOver(Request::input('P6'), ' ');
+            $cash->mon12 += (int)trimOver(Request::input('P7'), ' ');
 
             $cash->view_other_tills = Request::input('view_other_tills');
             $cash->network = $emp->network;
